feat: API 명세에 맞게 재구현
