+++ apps/client/src/pages/memo/apis/end-point.ts
@@ -1,7 +1,3 @@
 export const ALL_MEMO_END_POIINT = {
   GET: 'api/v1/memo',
 };
-
-export const TAG_END_POINT = {
-  GET: 'api/v1/tag',
-};
