+++ apps/client/src/pages/memo/apis/type.ts
@@ -2,6 +2,3 @@ import { paths } from '@shared/types/schema';
 
 export type AllMemoResponse =
   paths['/api/v1/memo']['get']['responses']['200']['content']['*/*'];
-
-export type TagResponse =
-  paths['/api/v1/tag']['get']['responses']['200']['content']['*/*'];
