+++ apps/client/src/shared/apis/tag/end-point.ts
@@ -1,0 +1,3 @@
+export const TAG_END_POINT = {
+  GET_TAG: 'api/v1/tag',
+};
