+++ apps/client/src/shared/apis/tag/query-key.ts
@@ -1,0 +1,4 @@
+export const TAG_KEY = {
+  ALL: ['user'],
+  GET_ALL: () => [...TAG_KEY.ALL, 'get'],
+};
